Rename and move parameters and util packages
