+++ geode-benchmarks/src/main/java/org/apache/geode/benchmark/parameters/JVMParameters.java
@@ -12,7 +12,7 @@
  * or implied. See the License for the specific language governing permissions and limitations under
  * the License.
  */
-package org.apache.geode.benchmark.tests.parameters;
+package org.apache.geode.benchmark.parameters;
 
 public class JVMParameters {
   public static final String[] JVM_ARGS = new String[] {
